Move function to cleaning utils
